requirements.txt
# Private Doc Q&A - Requirements
# Install: pip install -r requirements.txt

# Document parsing
pdfplumber>=0.10.0          # PDF extraction with tables
pypdf>=4.0.0                # Fallback PDF parser
python-docx>=1.1.0          # DOCX extraction

# Embeddings & Vector search
sentence-transformers>=2.2.0 # Local embeddings (MiniLM)
numpy>=1.24.0               # Vector operations

# Audio (optional - for voice input)
sounddevice>=0.4.6          # Audio recording

# Utilities
PyYAML>=6.0                 # YAML parsing for markdown frontmatter
